pyproject.toml
[project]
name = "borrowd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-allauth>=65.7.0",
    "django-browser-reload>=1.18.0",
    "django-cleanup>=9.0.0",
    # keep django-cotton locked because newer versions break debug mode -- check in again
    # towards the end of 2026 to see if we can upgrade again
    "django-cotton==2.1.2",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-guardian>=3.0.0",
    "pillow>=11.1.0",
    "psycopg[binary]>=3.1",
    "gunicorn>=23.0.0",
    "django-storages[s3]>=1.14.6",
    "django-vite>=3.1.0",
    "django-notifications-hq @ git+https://github.com/django-notifications/django-notifications.git",
    "django-imagekit>=6.0.0",
    "sentry-sdk[django]>=2.48.0",
]

[dependency-groups]
dev = [
    "pyyaml>=6.0.2",
]

[tool.djlint]
# https://djlint.com/docs/configuration/
profile = "django"
max_blank_lines = 1
line_break_after_multiline_tag = true
blank_line_after_tag = "load,extends,include"
custom_html = 'c-[\w.-]+' # cotton components
format_css = false
format_js = false
use_gitignore = true
